Refactor: Separate structure into folders/files
Also includes using a repository class for data structure
